+++ Makefile.am
@@ -126,7 +126,7 @@ src_libfabric_la_SOURCES =			\
 	include/fi_enosys.h			\
 	include/ofi_file.h			\
 	include/fi_indexer.h			\
-	include/fi_iov.h			\
+	include/ofi_iov.h			\
 	include/fi_list.h			\
 	include/fi_lock.h			\
 	include/fi_mem.h			\
