+++ libfabric.vcxproj.filters
@@ -479,6 +479,9 @@
     <ClInclude Include="include\ofi_file.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="include\ofi_iov.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="include\fi_lock.h">
       <Filter>Header Files</Filter>
     </ClInclude>
