+++ prov/sockets/src/sock_atomic.c
@@ -53,7 +53,7 @@
 
 #include "sock.h"
 #include "sock_util.h"
-#include <fi_iov.h>
+#include <ofi_iov.h>
 
 #define SOCK_LOG_DBG(...) _SOCK_LOG_DBG(FI_LOG_EP_DATA, __VA_ARGS__)
 #define SOCK_LOG_ERROR(...) _SOCK_LOG_ERROR(FI_LOG_EP_DATA, __VA_ARGS__)
