+++ prov/util/src/util_mr_cache.c
@@ -34,7 +34,7 @@
 #include <config.h>
 #include <stdlib.h>
 #include <fi_util.h>
-#include <fi_iov.h>
+#include <ofi_iov.h>
 #include <ofi_mr.h>
 #include <fi_list.h>
