+++ prov/verbs/src/ep_rdm/verbs_rdm_msg.c
@@ -31,7 +31,7 @@
  * SOFTWARE.
  */
 
-#include "fi_iov.h"
+#include "ofi_iov.h"
 #include "fi_enosys.h"
 
 #include "verbs_rdm.h"
