+++ prov/verbs/src/ep_rdm/verbs_tagged_ep_rdm.c
@@ -36,7 +36,7 @@
 
 #include <fi_list.h>
 #include <fi_enosys.h>
-#include <fi_iov.h>
+#include <ofi_iov.h>
 #include <rdma/fi_tagged.h>
 
 #include "verbs_queuing.h"
