+++ src/iov.c
@@ -35,7 +35,7 @@
 #include <string.h>
 
 #include <fi.h>
-#include <fi_iov.h>
+#include <ofi_iov.h>
 
 uint64_t ofi_copy_iov_buf(const struct iovec *iov, size_t iov_count, uint64_t iov_offset,
 			  void *buf, uint64_t bufsize, int dir)
